+++ SentryTestUtils/SentryTestUtils-ObjC-BridgingHeader.h
@@ -8,7 +8,6 @@
 
 #if SENTRY_HAS_UIKIT
 #    import "SentryAppStartTracker.h"
-#    import "SentryDisplayLinkWrapper.h"
 #    import "SentryFramesTracker+TestInit.h"
 #    import "SentryUIViewControllerPerformanceTracker.h"
 #endif // SENTRY_HAS_UIKIT
@@ -33,8 +32,8 @@
 #import "SentryCrashMonitor_MachException.h"
 #import "SentryDebugImageProvider+HybridSDKs.h"
 #import "SentryDependencyContainer.h"
-#import "SentryDispatchFactory.h"
 #import "SentryFileManager+Test.h"
+#import "SentryFileManagerHelper.h"
 #import "SentryHub+Private.h"
 #import "SentryHub+Test.h"
 #import "SentryLogC.h"