+++ Sources/Sentry/Processors/SentryWatchdogTerminationBreadcrumbProcessor.m
@@ -1,5 +1,4 @@
 #import "SentryWatchdogTerminationBreadcrumbProcessor.h"
-#import "SentryFileManager.h"
 #import "SentryInternalDefines.h"
 #import "SentryLogC.h"
 #import "SentrySerialization.h"