+++ Sources/Sentry/SentryCrashIntegrationSessionHandler.m
@@ -1,7 +1,6 @@
 #import "SentryCrashIntegrationSessionHandler.h"
 #import "SentryClient+Private.h"
 #import "SentryDependencyContainer.h"
-#import "SentryFileManager.h"
 #import "SentryHub.h"
 #import "SentryInternalDefines.h"
 #import "SentryLogC.h"