+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add SentryDistribution as Swift Package Manager target (#6149)
+
+## 8.56.2
+
+### Fixes
+
+- Fix crash from null UIApplication in SwiftUI apps (#6264)
+
 ## 8.56.1
 
 > [!Warning]